React Bootstrap - Section 4
Imported React Bootstrap and used the Card, Row, and Col display features to list the products on the HomeScreen page. Created a component for a single Product to display image and title on each card. Also created a component called Rating to display the rating and number of reviews on each card. Used a link from Font Awesome to show the ratings, stars, and half-stars.