Cart Screen - Section 6
Created React context to save shopping cart items when the 'Add to Cart' button is clicked. Added an id element to each product in the data table. Adding a unique item prevents the shopping cart from having duplicate items (and instead increases quantities). Created a new API to return products using the new id. Added new route for CartScreen to display the contents of the shopping cart as well as a button for checking out. Added navigation to direct user to the CartScreen after adding an item to the cart. Finally added some images to add, subtract and remove quantities from the shopping cart.